refactor: package structure rebuilding
