+++ src/main/java/com/flarecafe/feature/promotion/domain/evaluator/PeriodEvaluator.java
@@ -1,6 +1,6 @@
 package com.flarecafe.feature.promotion.domain.evaluator;
 
-import com.flarecafe.feature.promotion.domain.model.ConditionType;
+import com.flarecafe.feature.promotion.domain.support.ConditionType;
 import com.flarecafe.feature.promotion.domain.model.PromotionCondition;
 import com.flarecafe.feature.promotion.domain.PromotionConditionEvaluator;
