+++ src/main/java/com/flarecafe/feature/promotion/domain/model/PromotionCondition.java
@@ -3,6 +3,9 @@ package com.flarecafe.feature.promotion.domain.model;
 import com.flarecafe.feature.generic.BaseEntity;
 import com.flarecafe.feature.generic.Status;
 import com.flarecafe.feature.generic.TimeInterval;
+import com.flarecafe.feature.promotion.domain.support.ConditionType;
+import com.flarecafe.feature.promotion.domain.support.PromotionCategories;
+import com.flarecafe.feature.promotion.domain.support.PromotionMenus;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
