--- src/main/java/com/flarecafe/feature/promotion/domain/model/PromotionType.java
@@ -1,6 +1,0 @@
-package com.flarecafe.feature.promotion.domain.model;
-
-public enum PromotionType {
-  AMOUNT,
-  PERCENTAGE
-}
