+++ src/main/java/com/flarecafe/feature/promotion/domain/support/ConditionType.java
@@ -1,4 +1,4 @@
-package com.flarecafe.feature.promotion.domain.model;
+package com.flarecafe.feature.promotion.domain.support;
 
 public enum ConditionType {
   PERIOD,
