+++ src/main/java/com/flarecafe/feature/promotion/domain/support/PromotionMenus.java
@@ -1,5 +1,6 @@
-package com.flarecafe.feature.promotion.domain.model;
+package com.flarecafe.feature.promotion.domain.support;
 
+import com.flarecafe.feature.promotion.domain.model.PromotionMenu;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Embeddable;
 import jakarta.persistence.FetchType;
