+++ src/main/java/com/flarecafe/feature/promotion/domain/support/PromotionType.java
@@ -1,0 +1,6 @@
+package com.flarecafe.feature.promotion.domain.support;
+
+public enum PromotionType {
+  AMOUNT,
+  PERCENTAGE
+}
