+++ src/test/java/com/flarecafe/feature/promotion/domain/PromotionRepositoryTest.java
@@ -5,7 +5,7 @@ import com.flarecafe.config.JpaConfig;
 import com.flarecafe.feature.generic.Money;
 import com.flarecafe.feature.generic.Status;
 import com.flarecafe.feature.promotion.domain.model.Promotion;
-import com.flarecafe.feature.promotion.domain.model.PromotionType;
+import com.flarecafe.feature.promotion.domain.support.PromotionType;
 import com.flarecafe.feature.promotion.infra.PromotionJpaRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
